Smart-AI-Resume-Analyzer/requirements.txt
# Core Streamlit and Web Framework
streamlit
streamlit-option-menu
streamlit-lottie
streamlit-extras

# AI and Machine Learning
openai
google-generativeai
python-dotenv

# PDF and Document Processing (Cloud compatible - no compilation needed)
pdfplumber
pypdf
python-docx
pdfminer.six

# PyMuPDF for local development (original annotation system)
# May not work in Streamlit Cloud due to compilation requirements
PyMuPDF

# Data Processing (using stable versions with pre-built wheels)
pandas==2.0.3
numpy==1.24.3

# Database
sqlalchemy
openpyxl

# Document Generation
reportlab

# OCR and Image Processing (optional - may not work on all cloud platforms)
pdf2image
pytesseract

# Visualization
plotly

# Text Processing and NLP
nltk
spacy

# Image Processing
Pillow

# HTTP Requests  
requests

# Essential packages
typing-extensions

# Note: Some packages like selenium removed - not supported in Streamlit Cloud environment
# OCR packages (pdf2image, pytesseract) may not work without system dependencies
